requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-admin-interface==0.28.8
django-colorfield==0.11.0
django-grappelli==4.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
inflection==0.5.1
kombu==5.4.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.3.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
